+++ main.js
@@ -8,7 +8,7 @@
  */
 
 const jsonWriter = require('./json-writer')
-const xsdImport = require('./xsd-import')
+const istdBerichtXsdImport = require('./istd-bericht-xsd-import')
 
 const JSON_FILE_FILTERS = [
     { name: 'JSON Files', extensions: ['json'] },
@@ -44,7 +44,7 @@ function _handleIstdJsonExport(fullPath) {
 function _handleIstdBerichtImport(fullPath) {
     if (fullPath) {
         try {
-            xsdImport.importBerichtXsdFile(fullPath)
+            istdBerichtXsdImport.importBerichtXsdFile(fullPath)
         } catch (err) {
             console.error(err)
         }
@@ -52,7 +52,7 @@ function _handleIstdBerichtImport(fullPath) {
         var files = app.dialogs.showOpenDialog('Selecteer een Bericht (.xsd)', null, XSD_FILE_FILTERS)
         if (files && files.length > 0) {
             try {
-                xsdImport.importBerichtXsdFile(files[0])
+                istdBerichtXsdImport.importBerichtXsdFile(files[0])
             } catch (err) {
                 app.dialogs.showErrorDialog('Geen bestand kunnen selecteren.', err)
                 console.error(err)
